% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM10.R
\docType{data}
\name{jus}
\alias{jus}
\title{TITLE}
\format{The format is a list of 2 component:

$x: A data-frame containing input variables: with 100 obs. of 200
variables ;

$y: Output variable: a factor with 2 levels "-1" and "1".}
\source{
Weston, J., Elisseff, A., Schoelkopf, B., Tipping, M. (2003),
\emph{Use of the zero norm with linear models and Kernel methods},
J. Machine Learn. Res. 3, 1439-1461
}
\usage{
jus
}
\description{
DESCRIPTION
}
\examples{

##'
\dontrun{

}

}
\keyword{datasets}
