% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tune.R
\name{tune}
\alias{tune}
\alias{tune.VSURF_interp}
\alias{tune.VSURF_thres}
\title{Tuning of the thresholding and interpretation steps of VSURF}
\usage{
tune(x, ...)

\method{tune}{VSURF_thres}(x, nmin = 1, ...)

\method{tune}{VSURF_interp}(x, nsd = 1, ...)
}
\arguments{
\item{x}{An object of class \code{VSURF_thres} or \code{VSURF_interp}, which
is the result of the \code{\link{VSURF_thres}} or \code{\link{VSURF_interp}}
function.}

\item{nmin}{Number of times the "minimum value" is multiplied to set
threshold value. See details below.}

\item{nsd}{Number of times the standard deviation of the minimum value of
\code{err.interp} is multiplied. See details below.}

\item{\dots}{Not used.}
}
\value{
An object with the same structure than the original output (from
\code{\link{VSURF_thres}} or \code{\link{VSURF_interp}}).
}
\description{
This function allows to tune the "thresholding" and "interpretation step" of
VSURF, without rerunning all computations.
}
\details{
In \code{\link{VSURF_thres}} function, the actual threshold is performed
like this: only variables with a mean VI larger than \code{nmin} *
\code{min.thres} are kept.  The function \code{tune.VSURF_thres} allows you
to change the value of \code{nmin} (which multiply the estimated threshold
value \code{min.thres}), without rerunning all computations.  To get a
softer threshold than default, choose a value of \code{nmin} less than 1,
and to get a harder one, choose a value larger than 1.

In \code{\link{VSURF_interp}} function, the smallest model (and hence its
corresponding variables) having a mean OOB error rate less than
\code{err.min} + \code{nsd} * \code{sd.min} is selected.  The function
\code{tune.VSURF_interp} allows to change the value of \code{nsd} (which
multiply the standard deviation of the minimum OOB error rate
\code{sd.min}), without rerunning all computations.  To get a larger model
than default, choose a value of \code{nsd} less than 1, and to get a smaller
one, choose a value larger than 1.
}
\examples{
\dontrun{
data(iris)
iris.thres <- VSURF_thres(iris[,1:4], iris[,5], ntree = 100, nfor.thres = 20)
iris.thres.tuned <- tune(iris.thres, nmin = 10)
iris.thres.tuned
iris.interp <- VSURF_interp(iris[,1:4], iris[,5], vars = iris.thres$varselect.thres,
                            nfor.interp = 10)
iris.interp.tuned <- tune(iris.interp, nsd = 10)
iris.interp.tuned
}
}
\author{
Robin Genuer, Jean-Michel Poggi and Christine Tuleau-Malot
}
\references{
Genuer, R. and Poggi, J.M. and Tuleau-Malot, C. (2010),
\emph{Variable selection using random forests}, Pattern Recognition Letters
31(14), 2225-2236
}
\seealso{
\code{\link{VSURF}}, \code{\link{VSURF_thres}},
\code{\link{VSURF_interp}}
}

